+++ src/util/string-utility-functions.ts
